Added notifications for suggestions

+++ app/controllers/notifications_controller.rb
@@ -1,6 +1,6 @@
 class NotificationsController < ApplicationController
   before_action :logged_in_user, only: [:index, :important, :delete, :delete_all, :delete_all_important, :summary,
-                                        :summary_selection, :selection, :timeline, :reference, :comment,
+                                        :summary_selection, :selection, :timeline, :reference, :comment, :suggestions,
                                         :selection_win, :summary_selection_win, :selection_loss, :summary_selection_loss, :suggestion]
 
   def index
@@ -25,7 +25,11 @@ class NotificationsController < ApplicationController
                 if Notification.where(user_id: current_user.id, category: 6).count > 0
                   @filter = :selection
                 else
-                  @filter = :timeline
+                  if Notification.where(user_id: current_user.id, category: 7).count > 0
+                    @filter = :suggestions
+                  else
+                    @filter = :timeline
+                  end
                 end
               end
             end
@@ -57,6 +61,9 @@ class NotificationsController < ApplicationController
                                      :user_id).where(id: comment_ids).page(params[:page]).per(20)
       when :selection
         @selections = Notification.where(user_id: current_user.id, category: 6).page(params[:page]).per(20)
+      when :suggestions
+        suggestion_ids = Notification.where(user_id: current_user.id, category: 7).pluck(:suggestion_id)
+        @suggestions   = Suggestion.select(:id, :user_id, :comment ).where(id: suggestion_ids).page(params[:page]).per(20)
     end
   end
 
@@ -116,6 +123,13 @@ class NotificationsController < ApplicationController
         redirect_to notifications_index_path(filter: :selection)
         return
       end
+      if params[:notification][:suggestion_ids]
+        notifs = Notification.where(user_id:    current_user.id, category: 7,
+                                    suggestion_id: params[:notification][:suggestion_ids])
+        notifs.destroy_all
+        redirect_to notifications_index_path(filter: :suggestions)
+        return
+      end
     end
     redirect_to notifications_index_path
   end
@@ -134,6 +148,8 @@ class NotificationsController < ApplicationController
         Notification.where(user_id: current_user.id, category: 5).destroy_all
       when :selection.to_s
         Notification.where(user_id: current_user.id, category: 6).destroy_all
+      when :suggestions.to_s
+        Notification.where(user_id: current_user.id, category: 7).destroy_all
     end
     redirect_to notifications_index_path
   end
@@ -182,6 +198,12 @@ class NotificationsController < ApplicationController
                                  comment_id: notification_params,
                                  field:      field_params).destroy
     redirect_to comment_path(notification_params)
+  end
+
+  def suggestions
+    Notification.find_by(user_id:    current_user.id, category: 7,
+                         suggestion_id: notification_params).destroy
+    redirect_to suggestion_path(notification_params)
   end
 
   def selection_win
